LayoutTests/fast/canvas/canvas-transforms-fillRect-shadow-expected.txt

Ensure correct behavior of canvas with fillRect+shadow after translation+rotation+scaling. A blue and red checkered pattern should be displayed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is 255
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 127
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 127
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 127
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 106
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 106
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 83
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 36
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 36
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 36
PASS successfullyParsed is true

TEST COMPLETE
